Using Pascal Case styling (unscores) for variables
According to PEP8 [1] both functions and variable names should use the
underscore naming convention (ie, 'my_var') rather than camel case ('myVar').
This is a little weird for me and python standard libraries use both, but I see
a readability advantage to this for functions and my previous approach of using
camel case just for variables is kinda weird. Hence switching to the 'right'
convention while the codebase is still tiny.

As an added bonus this conforms with torctl...

[1] http://www.python.org/dev/peps/pep-0008/