Running tests based on input arguments
Making the test runner accept arguments for the type of tests to be ran. The
integration tests especially will take a while when they're implemented so
letting the user specify the use cases for those.

This included copying and refactoring some basic utilities from arm for
enumerations and terminal text attributes.