Moving integ runner into testing base
Moving the integ runner module from 'test/integ/runner.py' to 'test/runner.py'.
The 'test/unit/*' and 'test/integ/*' are for test cases, and will later be
kinda crowded. The runner is special since it doesn't test, but rather provides
the runtime context for the integ tests so probably best to keep this separate
from the rest.