Config argument for the test runner
Integration tests can have custom behavior via a testrc, which is simple for
now, and will grow as tests become more complex. Previously I was loading a
static settings.cfg but that was stupid. The user's config file should neither
be hardcoded nor under version control. This change also includes the config
loading in the status output to stdout.