Replacing get_entry with ControlLine class
Making a string subclass to help with the parsing of controller output. Most
entries are space separated lists of elements, which this class has functions
to easily parse. For controller messages that don't follow this pattern we can
still treat it as a normal string.

Next is to add header documentation and tests.