Util for reading proc contents
Rewrite of arm's utility for parsing proc contents, changes including better
error handling and refacotring for stem's coding conventions and commenting.

While writing the PROTOCOLINFO response handler I needed a system util for
querying tor's pwd, which works best with proc information, so pulling it all
in. For now most of this isn't being exercised, but will be later.