Standardizing on suppress_exc args for sys util
The system functions are mostly best-effort, platform specific attempts to
retrieve data and usually don't have meaningful information to provide back
to the caller via exceptions. I'm defaulting to have them return None, with
an optional arg for having them raise IOError exceptions on failure instead.

I've gone back and forth on the fencepost here a few times, and I'm still not
positive if this is the right choice. Might change this later if it doesn't
work out.