Running integ tests multiple connections methods
Adding a 'CONNECTION' target that, if set, will run integration tests with
multiple connection and authentication methods...
- no control connection
- control port with no auth
- control port with an authentication cookie
- control port with a password
- control port with both an authtication cookie and password
- control socket

This means running through the integ tests six times which currently results in
a runtime of arond fourty sectons, so this isn't the default.

The primary purpose for doing this is to exercise the PROTOCOLINFO parsing and
upcoming connecion methods with all of these tor configurations. The
ProtocolInfoResponse integ test doesn't yet actually test all of these - fixing
that is next.