Resetting system call mocking after test
The protocolinfo test mocked system calls but didn't reset the mock when it was
done. This didn't cause any errors but that was only luck (the system unit
tests probably ran afterward and cleared the mock when it was done). Oops, this
is gonna be an easy testing bug to introduce... :/