Skipping get_pid_by_name test when doomed
The get_pid_by_name function fails when there's multiple instances of a process
with that name. As a result the integ test for the function fails if there's
extra tor instances running on the system.

Using pgrep to check for other instances and skip those tests if they'd be
doomed to failure.