Adding an authentication exception hierarchy
For arm a substantial torctl pain point has been poor exception raising from
authentication functions. If everything raises a ValueError then the caller
needs to parse message responses or re-implement the code to get finer error
granularity.

Providing a hierarchy of authentication exceptions that can both make life
simple for callers, letting them catch the exceptions they want then fall
back to a catch-all for the AuthenticationFailure base.

One mildly icky point is that this relies on the error message provided from
tor to differentiate between when an authentication value is rejected verses
the type. That said...
- this issue is noted in the pydocs
- I have integ testing to check for error message changes
- callers are encouraged to use a general authenticate method which doesn't
  have this flaw since it uses the PROTOCOLINFO response (barring edge cases
  like a buggy PROTOCOLINFO response or the auth type changing in the split
  second between PROTOCOLINFO/AUTHENTICATE)
- even if the message changes direct callers will still be getting the right
  category of exception

Pydocs reference a general authenticate method but it's not included yet (first
draft is written but needs some work and testing).