Joining get_protocolinfo_* functions
Having a joined socket constructor and PROTOCOLINFO query doesn't make sense.
It allowed the user to make a PROTOCOLINFO query in a single line without a
leftover socket but that use case is so rare that it's fine for it to take a
couple more lines. This api is simpler and better for the common case.