Reconnecting socket after auth failures
Tor disconnects the control socket after Failed AUTHENTICATE calls in an effort
to mitigate the issue discussed in...
http://archives.seul.org/or/announce/Sep-2007/msg00000.html

This is unintuitive to stem users so I'm making a best effort attempt to
reconnect the socket in those cases. This isn't guaranteed to succeed, but
checking is_alive() is far nicer for callers than a try/connect/except block.