Integ test for general auth's password arg
Testing the various use cases for a password argument (missing argument,
incorrect password, and correct password). When we're running connection tests
with something other than password auth this is a very dull test (since we
succeed without raising a password exception), but when we just have a
HashedControlPassword this is a decent test.