General auth unit test for protocolinfo fetching
Unit tests for the general authentication function will be exercising it in
various error conditions by mocking out the functions it uses. This first test
exercises all of the use cases when the function needs to make a PROTOCOLINFO
query and, more importantly, lays the groundwork the rest of these unit tests
will use.