Unit test to exercise all general auth use cases
I'm aiming for both expansive and deep unit testing for the general
authenticate method and this does the former. It runs the authenticate()
function under all combinations of valid input and having authenticate_*
functions either succeed or raise one of its exceptions. At present this just
checks that the authenticate() function always either succeeds or raises a
AuthenticationFailure - an important test, but I'll next try to make it a
deeper validity check.