Moving null-handler logging hack into log util
Without a handler the builtin logging class emits a warning. To get around this
I added a no-op handler if none already existed before any logging was done via
a module import hack. However, now that we're using a log util we can put the
hack in that instead.

I tried to repro the warning but didn't have any success... maybe the issue
only exists for older python versions? Reguardless, this workaround doesn't do
any harm so keeping it around.