Downgrading protocolinfo version mismatch to INFO
The response for PROTOCOLINFO queries do not necessarily need to be what we
requested. We make a best effort to parse any response as a v1 response, and
warned if the response was for something besides v1. Downgrading the warning
from WARN to INFO since this is a valid tor response and, if we're still able
to parse it, the user doesn't really need to be alerted.

If the version mismatch can't be handled like a v1 response then that most
likely *will* cause warnings or errors that we'll alert the user of.