Integ for cookie auth failed due to missing cookie
If we're running the integ tests for the first time then it would fail because
it attempts cookie auth but the cookie file doesn't exist yet. This went
uncaught for quite a while because after the first test run we leave the data
directory (to speed up future test runs). Caught thanks to boerni.