Integ tests for stem.version
I didn't have any integ tests for versions? Seriously? *sigh*

Adding some basic verification tests that we can handle the version of our
system and test instances. Earlier boerni was running into a bug where another
integ tests inadvertantly failed due to an uncaught exception from
get_system_tor_version(). It was inappropriate for that test to fail (it wasn't
testing versions) but we should have some test for the function so this is it.