Mocking utility functions
As the tests grow I've found that I'm repeating a lot of mocking logic. Making
a nice, general purpose utility to make mocking easier and less error prone.

A benefit of this (and the original goal) is that we'll be able to remove
stem.util.system.CALL_MOCKING. Testing code in the library is icky, and once we
swap over the system integ tests we'll be able to remove this override
function.