Making lsof cwd checks account for permissions
With the anti-debugger tor change the /proc/<pid>/cwd symlink becomes only
readable by root. This causes lsof queries for a process' cwd to return the
proc path with a message saying that permission is denied, and the sys tools
misparsed that as meaning that the cwd _was_ the proc path.