Removing test.runner.get_connection_type()
Integration tests used the get_connection_type() method to either query or
infer attributes about the connection it was running against. This was stupid -
they should query the attributes directly and make test assertions based on
that. The get_connection_type() value was our testing target which should
simply be a user friendly tag for a set of testing attributes. Next I'll be
moving testing targets completely from the runner so they only exist in
'run_tests.py'.