Integ RUN_NONE target wasn't runnable
The RUN_NONE target wasn't being included among the runnable integration
targets because it had an empty torrc which failed a check that should have
been for the attribute's existance instead.

Also correcting the case of the 'TARGETS' enum. While it is a contant, I've
been going with class case contentions for enumerations.