Dropping Config.sync() in favor of config_dict()
The Config class' sync method was a step in the right direction, but the name
was confusing and the usage was suboptimal. In the vast majority of cases the
caller simply wants a dictionary that stays in sync with the configuration. The
config_dict() function is essentially the same as sync but with more succinct
calls.