Integ check for value of get_system_tor_version()
We had basic verification tests for get_system_tor_version() but it didn't
check the value. This was a refactoring mistake, iirc it once checked the value
but since we can run with an arbitrary tor binary I removed that, forgetting
that get_system_tor_version() could optionally take that binary path.

Making this a separate test since it relies on the tor instance being
accessible (which isn't the case with all targets).