Removing CALL_MOCKING from the system module
While writing the unit and integ tests I realized that the tests could be
greatly simplified by adding an override function called 'CALL_MOCKING'
that would let the tests modify the call function's behavior.

This was a mistake. Production code should be designed to make testing easy
because this leads to better production code, but adding hacks that are only
useful for testing is bad. Now that we have a nice mocking module the
CALL_MOCKING hack is no longer necessary so removing it.