Accounting for big-endian architectures in proc
The proc utils were assuming that encoded ip addresses were little-endian. This
was fixed in the project these utils are based on (psutil) and I'm adopting the
fix...
https://code.google.com/p/psutil/issues/detail?id=201
https://trac.torproject.org/projects/tor/ticket/4777

This evidently occures on OpenWRT (ar71xx), thanks to swalker for the catch!

This is the same as arm commit 8ec7095d79ecad9d3432193a2b1f9fdefab8d7f3