Exercising is_alive in ControlMessage integ tests
Adding a basic check for the is_alive method. The method is a little troubesome
because if you aren't continually pulling from the socket then you won't detect
a disconnect until the following usage.

This was already documented and iirc it won't be an issue in practice for
controllers since they continually pull from the socket. But still, for low
level controller interaction it's kinda confusing.