Separating BaseController into standalone class
The BaseController was previously a ControlSocket subclass because it was
easier to use when several of its methods were accessable. However, from an
implementation perspective the BaseController was a wrapper class, not a proper
subclass.

After experimenting a bit more I realized that I don't want the BaseController
to provide *all* of the ControlSocket methods. In particular, it doesn't make
sense for callers to use the send() and recv() when there will be a msg()
method similar to TorCtl's sendAndRecv(). The wrapper/subclass mix was also
just plain old confusing as an object-oriented design.

I'm moving the notifications up to the controller so the ControlSocket is very
similar to how it was a couple weeks ago. Also dropping the passthrough integ
tests since most of them will break without send/recv - I'll need to add some
more targeted tests for passthrough methods later.