Fixing another close() deadlock issue
The previous fix narrowed the window where close() / recv() calls could trigger
deadlock, but it didn't eliminate it. I'm adding another test that reliably
triggered deadlock in that case and narrowing the window even more (which fixed
the issue).

I'm a little worried that this doesn't completely eliminate the issue since
there is a theoretical race if recv() calls close after someone else calls
close() but before they set the boolean flag. That said, I'm not sure if this
is really an issue in practice.