Unit tests for the load_processed_files() function
Unit tests for stem.descriptor.reader's load_processed_files() function. This
mocks the open call to simulate file reads and exercise a variety of good and
malformed contents. This does not yet test for use cases where the file doesn't
exists or fails to be read due to permissions, but those will be integ tests.