Integ test for basic usage example
The header documentation contains two examples - the equivilant to 'hello
world' with this class and an example of making use of the high water mark the
DescriptorReader tracks. This adds an integ test for the first.

At present this integ test dies in a fiery... well, fire. This is because the
DescriptorReader isn't yet implemented. That's next...