Only provide absolute paths with get_processed_files()
The DescriptorReader's targets can be relative paths, and this is fine.
However, the load and save functions for processed_files expect absolute paths
so making get_processed_files() expand relative paths.

I'm on the fencepost about if we should do this or make the load/save functions
accept relative paths. Going with this until someone produces a use case where
this makes them sad.