Fixing sinister concurrency issue
Replacing the _is_reading event flag with a 'FINISHED' entry in the
_unreturned_descriptors queue. This is because python's queues stupidly have no
notion of flushing, so there's no method for me to make a reliable check of 'if
the reading thread is finished AND the queue is empty'. I may have called
'put'. I may have a proveably not-empty queue. But can I make that check work?
Nooooo. That is... frustrating. >:(