Notifying skip listeners of files that were already read
Skip listeners are supposed to be notified of all files that we skip, but
wasn't accounting for files that we skipped because their 'last modified'
timestamp indicated that they'd already been read.

Modifying the integ test for set_processed_files() to test this listener
condition (more graceful that the prior test that it had been doing).