Descriptor module with common functionality
Adding the descriptor module which has the Descriptor class (the common parent
for all descriptor types) and the function for parsing descriptors from a file.
This parser will later do type recognition to figure out how descriptors
should be parsed. This is identical to how metrics-lib does this.

Nice advantage is that we can now move all of the remaining 'TODO' notes out of
the reader. It's done. :)