Merging descriptor.py with module init
The syntax for using the descriptor module was stupid ("import
stem.descriptor.descriptor"). Using the json module as an example for using
__init__.py for base functions. Common ones will live in 'stem.descritor' but
the contents, like 'stem.descriptor.reader', are also accessable.

This is similar to how the os module works, though in that case they're using a
lot more python magic to also support multiple operating systems.