Resolving circular dependency with server_descriptor
... damn this was a pita to debug. A long explanation of the issue is:
http://stackoverflow.com/questions/8991520/python-conditional-module-object-has-no-attribute-error- with-personal-package

Short story is that the stem.descriptor module imported server_descriptor prior
to being fully initialized which caused a server_descriptor use of
stem.descriptor.Descriptor to fail.

The confusing was that it was actually reporting failure in using
"stem.descriptor" (the module itself) which made no sense because it was listed
within sys.modules. It turns out that the python importer uses module
attributes rather than sys.modules mappings to resolve statements (ie, it was
doing getattr(stem, "descriptor")). However, a module's added as an attribute
of its parent when its *finished* importing.

This fixes the unit tests, but integ tests are still getting stuck when they
use the new server_descriptor parsing.