Server descriptor test for metrics content
Test for parsing a single server descriptor and checking its content. This
includes a few fixes and improvements for the ServerDescriptorV2 class, the
most intersting of which is that declaring instance variables prior to init
makes a single reference. Hence all server descriptors were using the same exit
policy. Oops. :)