Allowing for extra args on lines with multiple
Well, crap. I've been developing against the version 2 spec on the assumption
that was both the majority use case and what my test instance was running.
However, that's evidently pretty ancient and I've been running against version
3 all along.

First change I'm spotting in the spec is that lines with multiple arguments
should ignore extras so doing that.