Integ test to validate our cached descriptor file
Parsing all of the descriptors in our cached descriptors file to see if it has
anything that'll make us choke (and by extension indicate either a tor or stem
bug).

This is an important yet lengthy test, taking around four second. Targets do
not effect this test so only running it once for the test run.