Including deprecated server descriptor attributes
I have mixed feelings about cluttering the server descriptor with dead fields,
but since we're already this close to being perfectly spec conformant we might
as well include this last bit. Adding the fields but logging an info level
warning if we see it, and checking in our cached descriptor that they do not
exist.