Skipping auth message check when cookie missing
When we first run our tests we don't have an authentication cookie unless we've
run with the RUN_COOKIE target. We have special handling for this, but it did
not account for running with an open connection (the default) in which case
we'd encounter a second exception while trying to check the failure message.