Fixes for prior version parsing fixes
I never cease to be amazed at how error prone something that sounds as simple
as 'parse the tor version' can be. Guess that's why we're writing a library...

This has a couple fixes for the prior set of fixes.

a. The runner was misparsing 'GETINFO version' output for verisons without a
   space.

b. Sinister bug (took me almost an hour to track down) where the results from
   our mocked "tor --version" call were cached and broke later tests. This only
   manifested if your tor version wasn't 0.2.2.35 and you ran both the unit and
   integ tests together.