Making system unit tests more platform independent
On Windows the alternate path separator caused the unit tests for expand_path()
to fail. Switching the expand_path() and is_relative_path() to not make any
effort to work on Windows and fixing the test to be platform independent. In
the future we should make these funtions work on Windows but this isn't vital
at the moment and a pita due to its '[A-Z]:\\' prefixes (I really don't wanna
do a regex just to check if a path is absolute...).