Splitting relay and bridge server descritors
Making separate classes for relay and bridge v3 server descriptors. This is the
more proper object model and allows us to easily handle their differences in
terms of parsing and validation.

This includes the implementation and an integ test, but the implementation
isn't yet complete (missing or-address parsing) and we still need unit tests.
The current unit tests pass but the integ test fails due to or-address.