Differentiating read bridge and relay descriptors
The last integ failure was due to our attempt to read a bridge descriptor file
as a relay descriptor, which doesn't work because it is missing cryptographic
entries. Adding a bit of logic to differentiate relay from bridge descriptors
so we can parse them as their appropriate type. The cues the difference based
on the 'Unnamed' nickname and '10.x.x.x' address scrubbing.