Removing deprecation warnings for server descriptor fields
The read/write-history and eventdns server descriptor fields are part of the
spec but no longer used. I was logging a notice that they're deprecated when
we encountered them, but this is neither useful nor does it make sense. We
might be parsing old metrics archives where those attributes are perfectly
valid.

Integ tests include a check that those aren't seen in the wild, which is what I
really care about.