Removing timeout from descriptor reader enqueue calls
The reader had a timeout for enqueue operations so it could periodically check
if we were stopped while waiting to insert a descriptor that we've read. As
Karsten thought we were able to do better. Dropped the timeout by clearing
our queue on shutdown (to unblock any enqueue calls in process), and skip
further enqueuing while we're shutting down.

Both iterating and stopping are under a read lock so we don't need to worry
about this changing the order in which descriptors are provided to callers.