Allowing for negative uptimes prior to tor 0.1.2.7
Tor 0.1.2.7 fixed a bug which could result in negative uptime values. Making
our descriptor parser accept negative uptimes for tor versions with the bug,
and fail validation for later versions.

This includes a test with one of the problematic descriptors, checking that
we can both parse it, and will fail validation if it was for a later version.

Thanks to Karsten for figuring out where these values were coming from!