Fixing testing thread leak and adding check for it
Python does not gracefully handle lingering threads at shutdown, even if
they're daemons (... which is dumb since that's the only point of the daemon
thread). Having threads causes occasional nonsensical stacktraces with things
like "'NoneType' object has no attribute 'socket'" for module references.

Adding a check after running our integ tests that only the main thread remains,
and erroring if that's not the case. This also fixes the thread leak that was
probably the issue in...
https://trac.torproject.org/projects/tor/ticket/5512