Concurrency is still hard
It felt like the handle_close flag was hacky and now I know why. If recv()
tried to call close() while a send() call was in flight and would also
eventually call close() then we'd encounter a similer type of deadlock as
commit d5162f4 tried to fix.

This did not come up for control port connections, but it did frequently occure
for control sockets (as seen by deadlock when running the RUN_SOCKET target).
This did not always occure because the recv() sometimes acquired the send_lock
before the next send() call, making everything proceed correctly.

Fixing this by having recv() make a non-blocking attempt to get the send_lock.
If it works then we know that we can safely close, and if we can't acquire the
lock then we know that a send() or close() call is already in progress and can
leave the closing to them.

This does introduce a possible issue where the send() call succeeds, the recv()
call fails with a SocketClosed, and we're left in a state of being alive.
However, this is both a weird use case (how can the send() work if the socket
is closed?) and also not strictly wrong (raising a SocketClosed does not mean
that we've finished shutting ourselves down). It's a little tempting to add a
dedicated close_lock to account for this, but after looking into that I
realized that this would make the concurrency far more convoluted.