Separating bridge scrubbing validation into method
Bridge descriptor scrubbing is a moving target and it's quite possible that
we'll need to parse descriptors that conform to a newer or older scheme. Hence
making scrubbing validation part of the constructor is inappropriate. Moving it
to a separate method instead.

This includes two methods, one to simply check if we think that the scrubbing
is right, and another to get descriptions of the issues. This is of limited
usefulness to callers, so we might need to switch to an exception hierchy
later. However, we don't have any use cases that care to check the scrubbing
yet so leaving that alone for now.