Making descriptor reader persistence more convenient
Our current functions for loading/saving processed file listings are fine if
you want error handling and a great deal of control. However, I suspect that
most callers would prefer for this to be an attribute of the reader itself.

Adding an argument that performs best-effort persistance of our processed files
listing.