Skeleton for extra-info descriptors
Basic module for parsing extrainfo descriptors. This doesn't actually do any
parsing yet, just turning the raw content into a ExtraInfoDescriptor instance.
This abstracts code we'll need out of the server_descriptor module so we'll be
able to use it.

What code there is here is exercised via the runner tests.