Expanding launch_tor to support commandline options
When writing small scripts we often want to start a tor instance for a special
purpose, for instance to test a new feature. When doing this we want a
bare-bones tor instance with just a few specific options. Adding better support
for this by making the launch_tor function accept command-line options and the
ability to use a blank torrc.