Integ tests for launch_tor
I've been leaving launch_tor untested because it was exercised in order to run
the integ tests and any tests involving it would take a long time, on the order
of a dozen seconds! Hey, it's a long time when you run this as often as I do...

It's an important function and has more options than what's exercised for
running integ tests so adding some tests for it.