Reorganizing extrainfo fields by value type
Grouping both parsing and unit tests by the value format to make it more
maintainable and easier to add the dozens of other attributes. Also using a
'dir_' prefix for all directory mirroring attributes rather than 'dirreq_' (I
was previously being inconsistent on this).

This expands the testing to cover more attributes, along with adding support
for the "dirreq-v2-ips" and "dirreq-v3-ips" lines. Those lines revealed some
misses in how I was parsing the maxmind locales. In particular I was missing
'??' and numeric locales like 'a1' which it uses for an 'Anonymous Proxy'.