Replacing enums with class instances for connect_* functions
The connect_port and connect_socket_file convenience functions accept a
controller type that we would like to return. I had been using an enum for this
argument, but this means that we both (a) need to manually update it when
adding controllers and (b) can't recognize controller the user makes.

Replacing this with a class instance, like what TorCtl does. Tested by running
against a script for printing BW events.