Adding separate funtion for launching tor with tmp torrc
Like Vidalia, adding a function that creates a torrc, launches tor with it,
then removes it from disk. This is preferable to using the commandline
arguments since the those shouldn't be used for substantial data nor the hashed
password.

This replaces launch_tor()'s options argument, though I added a more genral
args replacement that can be used to do the same if the user wants.