Supporting bridge descriptor's router-digest field
Scrubbed bridge server and extra-info descriptors will soon have a
'router-digest' line which contains the server digest. This is a value that we
can't calculate for ourselves because, of course, the descriptor's contents
have been modified.

I'm flagging this as being a required field. This does not yet parse it in
extra-info descriptors since it neither has a digest() method nor did I realize
that bridge extra-info descriptors were special.