Validate that signing key hash matches fingerprint
When parsing server descriptors checking that their fingerprints match a hash
of their signing key as part of validation. This requires the rsa module which
is neither built in, nor is it packaged for debian distros. Installation of it
is easy, but requires pip or easy-install. Instructions are available at...
http://stuvel.eu/files/python-rsa-doc/installation.html

It looks like python's builtin crypto module might be capable of doing this as
well...
http://stackoverflow.com/questions/5000434/python-rsa-library
https://www.dlitz.net/software/pycrypto/api/current/Crypto.PublicKey.RSA._RSAobj-class.html#exportKey

However, those instructions include usage of a private member and the function
that they suggest dosn't exist on my system (python 2.7.1), so I feel pretty
justified in saying "the pycrypto builtin is crap for this use case, and
patches welcome if someone can figure out how to make it work".

All credit for this patch goes to Beck, who's been diving into the descriptor
crypto on...
https://trac.torproject.org/projects/tor/ticket/5810