Generating sphinx docs for all stem modules
Using the sphinx-apidoc command to generate documentation for all stem modules.
I'm doing this via the make file to simplify the documentation creation process
(you still just need to run 'make html'). Also removing the windows make.bat
file since I'm not gonna be keeping it up to date. We can resurrect it if
someone volunteers to maintain it.