Targeting contents at top of modules
For modules like stem.response and stem.util I was linking to the first anchor,
which isn't the behavor I wanted (skips the title at the top, which is a nicer
landing). I resisted making this change earlier because it means that those
links have different styling from those above, but now that there's more of
them I'm happy with it this way.