Targeting reader test at /usr rather than root dir
I've been running the reader's test_stop() integ test by running over the root
direcctory because I simply needed something 'big' to keep the test occupied
for a second or two. This turned out to inadvertantly fuzz the reader which is
sorta a good thing since it discovered that block devices and encrypted
partitions can be slow to read, causing stop() to block for a bit.

However, we don't want to run over people's private stuff, and slow reads on
encrypted partitions probably isn't a problem that we want to solve anyway, so
running over '/usr' instead.