Adding future 'with' import and checking for it
Stem aims for python 2.5 compatability, but in that version the 'with' keyword
wasn't available by default and would only function with a 'future' import.
This has the possability of being a continuing pain in our ass so adding this
as a check to the check_whitespace module. Admittedly this isn't a whitespace
issue, but that module is just so damn conveninent for lintian style issues
too...

While doing this I discovered that we aren't running the checker over the stem
module (just the tests, it seems), so something else to fix...