Adding --test argument to selectively run tests
Adding an argument that'll only run tests matching the given module prefix.
This is to make it faster to write new tests (so we don't need to wait a minute
for the full integ suite to run each time).

Also adding examples to the '--help' output and rearranging the options so
'--config' is at the end (it isn't especially useful and may be dropped in the
future.