Dumbing down syntax for python 2.5
Evidently even with the 'with' keyword's future import it still chokes on comma
delimited entries (instead you need to do nested 'with' blocks). Ick. Oh well,
if that's all we need to maintain 2.5 compatability then I'll be happy.

It's a pity we can't tell python 'run in 2.5 compatability mode' to check for
these things. I doubt that we'll remember to run against a 2.5 instance to keep
stem working there. Oh well, guess this is a problem that'll need to be
addressed when figuring out a release procedure.