The followlinks arg of os.walk needs python 2.6
We were using the 'followlinks' argument two places: in the descriptor reader
and whitespace checker. For the former I'm noting that 2.6 is a requirement for
the argument and coding around it to work on 2.5. For the later I don't really
care if the whitespace checker follows links so I'm simply dropping the arg.