Adding todo notes around the stupid concessions to 2.5
I hate dirtying our code for ye old python 2.5. It's ancient and that fact
isn't getting rosier very fast. However, a lot of people still use it so we
should still support it for now. This won't always be the case so including
TODO notes to remove these icky bits when we drop compatability for it.