TestCase's skipTest method unavailable before python 2.7
Blarg! The unit testing's skipTest method was a new addition in python 2.7.
This is especially pesky since skipTest alters the control flow of the tests
(raising an exception so we can place it in setUp methods).

Options are to either monkey patch support into our testing module (ick!) or
add a fallback everywhere we use skipTest. This change adds a best-effort skip
method to the runner, then leaves it to the caller to return out of the method
(reporting success for skipped tests).

Sucks, but best that we can do until we drop *both* python 2.5 and 2.6
compatability, which probably won't happen until we finally move to the 3.x
series.