Integ test relied on tor being in path
I thought that it was a safe assumption that tor was in our path since the
integ tests will only run if it can start tor. However, the tests have a
'--tor <path to tor>' argument which means that the tests may be running
without our path including the tor binary (or even having the binary called
'tor').

Changing the test to look for 'ls' or 'dir' instead since every platform that I
can think of has that. Caught by Karsten.