Respecting reader stop() calls in mid-directory
If we, say, attempted to read a directory with a million files then the
reader's stop() method won't be respected until we've finished with that
directory. That's stupid. Checking the _is_stopped Event after processing
each file.

Python lacks targeted breaks, so I needed to refactor the code a bit to
accomidate this.