Revisions for prior SETCONF and RESETCONF additions
Several revisions to the prior changes, the most prominent of which is to use
an API more similar to arm's. The SETCONF and RESETCONF methods are
functionally identical except how they handle undefined values, so there's
little reason to provide separate methods in our controller. Just using an
optional 'reset' arg instead.

One thing that arm got wrong (and so did stem) is that our method to change
multiple options should accept a dict rather than a tuple list. Order *might*
matter for the accursed HiddenService options, but for the 99.9% use case a
tuple list is a very strange argument to accept for this.

This also includes some testing fixes, for instance...
* Testing when we have a list argument (for instance, setting our ExitPolicy).
* Failing our test if one of the calls that should fail actually succeeds.
* Reverting the configuration options that we change after the test.
* An identical test for context sensitive options was added to both the getconf
  and setconf test. Definitely a good use case to check and it could belong
  with either setconf or getconf, but doing the same test twice is pointless.